add editorconfig and comments for command provider
