+++ .editorconfig
@@ -1,0 +1,30 @@
+# Manage common editor configurations
+# set for example indentation tab and space
+# VS Extension: https://visualstudiogallery.msdn.microsoft.com/c8bccfe2-650c-4b42-bc5c-845e21f96328
+# more extensions and plugins for your favorite editor here: http://editorconfig.org/#download
+
+; Top-most EditorConfig file
+root = true
+
+; 4-column space indentation
+[*]
+indent_size = 4
+indent_style = space
+insert_final_newline = true
+trim_trailing_whitespace = true
+
+[*.md]
+indent_size = 4
+indent_style = space
+trim_trailing_whitespace = false
+
+[*.json]
+indent_size = 2
+indent_style = space
+trim_trailing_whitespace = false
+
+[*.min.*]
+max_line_length = off
+indent_style = ignore
+trim_trailing_whitespace = false
+insert_final_newline = ignore
