+++ commands-extender/commands-provider.ts
@@ -2,8 +2,16 @@ import { CommandProvider, CommandsData, Command, COMMANDS_TOKEN, CommandsTarget,
 import { Observable } from "rxjs";
 import { ClassProvider, Injectable } from "@angular/core";
 import { Router } from "@angular/router";
+import { PrintPreviewCommand } from "./print-preview.command";
 
+/**
+ * The category name in which to place the custom commands.
+ */
 const CUSTOM_CATEGORY_NAME = "Custom";
+
+/**
+ * The command model containing the metadata of the command.
+ */
 const CUSTOM_COMMAND_BASE: CommandModel = {
     name: "Custom",
     title: "Print preview",
@@ -11,18 +19,39 @@ const CUSTOM_COMMAND_BASE: CommandModel = {
     category: CUSTOM_CATEGORY_NAME
 };
 
+/**
+ * The category model.
+ */
 const CUSTOM_CATEGORY: CommandCategory = {
-    name: "Custom",
+    name: CUSTOM_CATEGORY_NAME,
     title: "Custom commands"
 };
 
+/**
+ * The comamnd provider provides the necessary commands back to the AdminApp.
+ */
+@Injectable()
 class DynamicItemIndexCommandProvider implements CommandProvider {
+
+    /**
+     * The method that gets invoked asking for the command models when the action menu is constructed.
+     * @param data The data needed to detirmine the types of command to return
+     * and where to place them - in the list or in edit mode
+     */
     getCommands(data: CommandsData): Observable<CommandModel[]> {
+
+        // the commands to be returned
         const commands = [];
+
+        // we wish to inject this command only in the list view
         if (data.target === CommandsTarget.List && data.dataItem) {
             const previewCommand = Object.assign({}, CUSTOM_COMMAND_BASE);
+
+            // assign an injection token or just the class
             previewCommand.token = {
                 type: PrintPreviewCommand,
+
+                // assign custom properties to be passed in the context
                 properties: {
                     dataItem: data.dataItem
                 }
@@ -31,31 +60,24 @@ class DynamicItemIndexCommandProvider implements CommandProvider {
             commands.push(previewCommand);
         }
 
+        // return an observable here, because this might be a time consuming operation
         return Observable.of(commands);
     }
 
+    /**
+     * The method that gets invoked asking for the category models when the action menu is constructed.
+     * @param data The data needed to detirmine the types of command to return
+     * and where to place them - in the list or in edit mode
+     */
     getCategories(data: CommandsData): Observable<CommandCategory[]> {
         return Observable.of([CUSTOM_CATEGORY]);
     }
 }
 
-@Injectable()
-export class PrintPreviewCommand implements Command {   
-    constructor(protected router: Router) {}
-
-    execute(context: ExecutionContext): Observable<any> {
-        const dataItem = context.data.dataItem;
-        const currentQueryParams = {
-            entitySet: dataItem.metadata.setName,
-            id: dataItem.key,
-            provider: dataItem.provider
-        }
-        const url = `/print-preview`;
-        this.router.navigate([url], { queryParams: currentQueryParams });
-        return Observable.of(null);
-    }
-}
-
+/**
+ * Export a 'multi' class provider so that multiple instances of the same provider can coexist.
+ * This allows for more than one provider to be registered within one or more bundles.
+ */
 export const COMMANDS_PROVIDER: ClassProvider = {
     useClass: DynamicItemIndexCommandProvider,
     multi: true,
