+++ commands-extender/print-preview.component.ts
@@ -2,18 +2,36 @@ import { Component, OnInit } from "@angular/core";
 import { ActivatedRoute } from "@angular/router";
 import { HttpClient } from "@angular/common/http";
 import { HTTP_PREFIX } from "progress-sitefinity-adminapp-sdk/app/api/v1";
+
+/**
+ * A component that loads the data item from the OData rest services 
+ * and display's the title of the data item.
+ */
 @Component({
     template: require("./print-preview.component.html")
 })
 export class PrintPreviewComponent implements OnInit {
+
+    /**
+     * The data item used in the view.
+     */
     protected dataItem: any;
 
+    /**
+     * Initializes a new instance of the PrintPreviewComponent
+     * @param http The http client service used for making http requests.
+     * @param route The current navigated route.
+     */
     constructor(
         private http: HttpClient,
         private route: ActivatedRoute) {
     }
 
+    /**
+     * NgOnInit lifecycle hook. Attached here so we can fetch the data on initialization
+     */
     ngOnInit() {
+        // get the route params that contain the metadata needed to load the data item
         const routeParams = this.route.snapshot.queryParams;
         
         // http prefix is dynamically replaced with the actual url of sitefinity
